Move db tests to different file and check for error on delete
